src_v3/Reflections_v3/requirements.txt
fastapi
fasthtml
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-ai